+++ src/gort/overwatcher/observer.py
@@ -15,11 +15,17 @@ from typing import TYPE_CHECKING
 
 from lvmopstools.retrier import Retrier
 
-from gort.exceptions import GortError, OverwatcherError, TroubleshooterTimeoutError
+from gort.exceptions import (
+    GortError,
+    OverwatcherError,
+    TroubleshooterCriticalError,
+    TroubleshooterTimeoutError,
+)
 from gort.exposure import Exposure
 from gort.overwatcher import OverwatcherModule
 from gort.overwatcher.core import OverwatcherModuleTask
 from gort.overwatcher.transparency import TransparencyQuality
+from gort.overwatcher.troubleshooter.recipes import AcquisitionFailedRecipe
 from gort.tile import Tile
 from gort.tools import cancel_task, decap, ensure_period, run_in_executor
 
@@ -193,6 +199,17 @@ class ObserverOverwatcher(OverwatcherModule):
 
         # If immediate, cancel the task now and cleanup.
         if immediate:
+            if self._starting_observations:
+                # Just let the startup finish. observe_loop_task() will immediately
+                # return because we have set _is_cancelling=True.
+                await self.notify("Waiting for the start-up sequence to finish.")
+                if block:
+                    while True:
+                        if not self._starting_observations:
+                            break
+                        await asyncio.sleep(1)
+                return
+
             self.observe_loop = await cancel_task(self.observe_loop)
 
             # The guiders may have been left running or the spectrograph may still
@@ -206,6 +223,11 @@ class ObserverOverwatcher(OverwatcherModule):
 
     async def observe_loop_task(self):
         """Runs the observing loop."""
+
+        # Immediately return if we are cancelling. This can happen if we disable
+        # the overwatcher during startup.
+        if self.is_cancelling:
+            return
 
         # Check that we have not started too early (this happens when we open the dome
         # a bit early to avoid wasting time). If so, wait until we are properly in
@@ -238,6 +260,9 @@ class ObserverOverwatcher(OverwatcherModule):
                     f"Received tile {tile.tile_id} from scheduler: "
                     f"observing dither positions {tile.dither_positions}."
                 )
+
+                if self.is_cancelling:
+                    break
 
                 await self.check_focus(force=n_tile_positions == 0 or self.force_focus)
 
@@ -397,6 +422,23 @@ class ObserverOverwatcher(OverwatcherModule):
                     )
 
             await self.gort.specs.reset()
+
+        # Use the code in the troubleshooter to check if all AG cameras are connected.
+        acq_failed_recipe = AcquisitionFailedRecipe(self.overwatcher.troubleshooter)
+
+        self.log.debug("Checking AG cameras.")
+        ag_pings = await acq_failed_recipe.ping_ag_cameras()
+
+        if not all(ag_pings.values()):
+            self.log.error("Not all AG cameras ping. Running troubleshooting recipe.")
+            try:
+                await acq_failed_recipe._handle_disconnected_cameras()
+            except TroubleshooterCriticalError as err:
+                await self.notify(
+                    f"Critical error while checking AG cameras: {decap(err)}",
+                    level="error",
+                )
+                return False
 
         return True